Module reorg per PR discussion
