+++ moneybot/markets/adapters/__init__.py
@@ -7,8 +7,8 @@ from typing import Generator
 from typing import Iterator
 from typing import List
 
-from moneybot.markets.proposed_trade import ProposedTrade
 from moneybot.markets.state import MarketState
+from moneybot.strategy import ProposedTrade
 
 
 logger = getLogger(__name__)
