+++ moneybot/markets/adapters/live.py
@@ -9,8 +9,8 @@ from typing import Iterator
 from poloniex import Poloniex
 
 from moneybot.markets.adapters import MarketAdapter
-from moneybot.markets.proposed_trade import ProposedTrade
 from moneybot.markets.state import MarketState
+from moneybot.strategy import ProposedTrade
 
 
 logger = getLogger(__name__)
