+++ tests/test_basic.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 import pytest
 
+from moneybot.examples.strategies import BuffedCoinStrategy
+from moneybot.examples.strategies import BuyHoldStrategy
+from moneybot.examples.strategies import PeakRiderStrategy
 from moneybot.fund import Fund
 from moneybot.markets.adapters.backtest import BacktestMarketAdapter
-from moneybot.strategies.buffed_coin import BuffedCoinStrategy
-from moneybot.strategies.buy_hold import BuyHoldStrategy
-from moneybot.strategies.peak_rider import PeakRiderStrategy
 
 
 @pytest.fixture
